user_service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = ""
authors = [
    {name = "mtr8650",email = "mohamadtaha8650@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11 , < 4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "beanie (>=1.29.0,<2.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
